- rename RTL module to getEffectiveRTL
